Fix clang-tidy warnings across main/: type safety, reserved identifiers, loop variables
- Reserved identifiers (_ltrim/_rtrim/_trim, _encoding_*, _response_404)
  renamed to avoid UB in the global namespace
- Narrowing conversions made explicit with static_cast<> throughout:
  atoi()→int8_t, float→uint16_t, size_t→int, uint32_t→float, etc.
- uint8_t loop variables upgraded to size_t where bound is sizeof()
- bugprone-undelegated-constructor: SSDP.h String ctors now use
  delegating-constructor syntax instead of creating discarded temporaries
- bugprone-signed-char-misuse: int8_t pin comparisons against uint8_t,
  RSSI widening, error-code widening made explicit
- performance-unnecessary-value-param: IPAddress and WiFiClient params
  changed to const references in Web.cpp/Web.h and SSDP.cpp/SSDP.h
- switch statements given explicit default: break where fall-through was
  implicit
- .clang-tidy updated to suppress noise (modernize-*, readability-*)
  and focus on bugprone-* and performance-*

+++ main/ControllerNetwork.cpp
@@ -202,7 +202,7 @@ void ControllerNetwork::emitSockets(uint8_t num) {
         json->addElem("channel", (int32_t)this->channel);
         json->endObject();
         sockEmit.endEmit(num);
-        this->lastRSSI = WiFi.RSSI();
+        this->lastRSSI = static_cast<int>(WiFi.RSSI());
         this->lastChannel = WiFi.channel();
       }
       else {
@@ -240,7 +240,7 @@ void ControllerNetwork::setConnected(conn_types_t connType) {
     this->_connecting = false;
     this->ssid = WiFi.SSID();
     this->mac = WiFi.BSSIDstr();
-    this->strength = WiFi.RSSI();
+    this->strength = static_cast<int>(WiFi.RSSI());
     this->channel = WiFi.channel();
     this->connectAttempts++;
     esp_wifi_set_ps(WIFI_PS_NONE);  // Disable modem sleep to prevent DELBA/frame-drop issues
